+++ examples/lookup_id.rs
@@ -47,7 +47,7 @@ fn main() {
     let api = ApiBuilder::new(from, secret).into_simple();
     match api.lookup_id(&criterion) {
         Err(e) => {
-            println!("Could not look up id: {:?}", e);
+            println!("Could not look up id: {}", e);
             process::exit(1);
         }
         Ok(id) => println!("The id is {}", id),