+++ examples/send_e2e_file.rs
@@ -127,6 +127,6 @@ fn main() {
     let msg_id = api.send(&to, &encrypted, false);
     match msg_id {
         Ok(id) => println!("Sent. Message id is {}.", id),
-        Err(e) => println!("Could not send message: {:?}", e),
+        Err(e) => println!("Could not send message: {}", e),
     }
 }